Release 4.39 - Gen
Change background hue of the icon display to reflect air quality levels. Add a disable display config option to only display the unit's serial number on a background hue that represents the air quality levels and change the temperature and humidity compensation when the weather cover is used. Add a reboot via the watchdog if comms is lost for an extended time and both Adafruit IO and Luftdaten are enabled. Only display indoor data on an indoor unit that has indoor/outdoor functionality  enabled and comms with the outdoor unit is lost.